requirements.txt
google-cloud-bigquery==3.40.1
pandas==2.3.3
matplotlib==3.10.8
seaborn==0.13.2
ipywidgets==8.1.8
streamlit==1.55.0
plotly==6.6.0
